Lazy way to print last line
Currently, when a signal is sent, all threads stop what they're doing
and die. Thread 0, which has been testing the lower range of numbers
prints out the maximum number it got to and the number of primes it
found. Thus the work done by other threads is sometimes discarded. Extra
IPC can mitigate this, but the only way to get rid of this effect
totally requires either (1) doing a large computation *after* the
timeout kill signal or (2) passing a ton of messages to verify primes
are tested in order.
There's probably a middle path, but I'm not exactly sure what it is.
Anyway, this gets 75/75 on submitty